+++ NEWS.md
@@ -2,6 +2,10 @@
 
 * This release of the R package builds against [TileDB 2.17.1](https://github.com/TileDB-Inc/TileDB/releases/tag/2.17.1), and has also been tested against earlier releases as well as the development version (#593)
 
+## Improvements
+
+* Set conditions are supported in query condition expressions (#597)
+
 ## Bug Fixes
 
 * The DESCRIPTION file now correctly refers to macOS 10.14 (#596)